requirements.txt
langgraph>=0.4.0
langchain-core>=0.3.0
google-genai>=1.0.0
requests>=2.31.0
streamlit>=1.35.0
python-dotenv>=1.0.0
pydantic>=2.7.0
pillow>=10.3.0
